COMMONS-479 : code refactoring
